Unregister local server
